HACK: Excluded JPEG format from facets
Hard coded an exclusion that prevents JPEG format from showing up in
facets. This is because the way we use JPEG resources for Application
screenshots was causing a ton of JPEG formatted resources to show up in
the facet even though they are not technically resources we want people
to be able to find.

A more robust solution to this would be to add exclusion list config
variables to ckan.ini and reference those for the exclusion in the
package_search function where I currently put the JPEG hack.